Fix focus issue on array pages
