+++ src/applications/appeals/shared/utils/focus.js
@@ -12,7 +12,7 @@ import { $, $$ } from '~/platform/forms-system/src/js/utilities/ui';
 
 import { LAST_ISSUE } from '../constants';
 
-export const focusFirstError = (root = document) => {
+export const focusFirstError = (_index, root) => {
   const error = $('[error]', root);
   if (error) {
     scrollToFirstError({ focusOnAlertRole: true });
@@ -21,27 +21,26 @@ export const focusFirstError = (root = document) => {
   return false;
 };
 
-export const focusEvidence = (_index, root) => {
+export const focusEvidence = (index, root) => {
   setTimeout(() => {
-    if (!focusFirstError(root)) {
+    if (!focusFirstError(index, root)) {
       scrollTo('topContentElement');
       focusElement('#main h3', null, root);
     }
   });
 };
 
-export const focusH3AfterAlert = ({
-  name,
-  onReviewPage,
-  root = document,
-} = {}) => {
+export const focusH3AfterAlert = (
+  index,
+  { name, onReviewPage, root = document } = {},
+) => {
   if (name && onReviewPage) {
     focusReview(
       name, // name of scroll element
       true, // review accordion in edit mode
       true, // reviewEditFocusOnHeaders setting from form/config.js
     );
-  } else if (!focusFirstError(root)) {
+  } else if (!focusFirstError(index, root)) {
     scrollTo('topContentElement');
     focusElement('h3#header', null, root);
   }
@@ -140,21 +139,20 @@ export const focusRadioH3 = () => {
 };
 
 // Testing focus on role="alert" inside web components
-export const focusH3OrRadioError = () => {
+export const focusH3OrRadioError = (_index, root) => {
   scrollTo('topContentElement');
-  const radio = $('va-radio, va-checkbox-group');
+  const radio = $('va-radio, va-checkbox-group', root);
   const hasError = radio.getAttribute('error');
   const target = hasError ? '[role="alert"]' : 'h3';
-  const root = hasError ? radio.shadowRoot : document;
-  waitForRenderThenFocus(target, root);
+  waitForRenderThenFocus(target, hasError ? radio.shadowRoot : root);
 };
 
 // Temporary focus function for HLR homlessness question (page header is
 // dynamic); once 100% released, change homeless form config to use
 // `scrollAndFocusTarget: focusH3`
-export const focusToggledHeader = () => {
+export const focusToggledHeader = (_index, root) => {
   scrollTo('topContentElement');
-  const radio = $('va-radio');
+  const radio = $('va-radio', root);
   if ((sessionStorage.getItem('hlrUpdated') || 'false') === 'false' && radio) {
     waitForRenderThenFocus('h3', radio.shadowRoot);
   } else {
@@ -162,14 +160,14 @@ export const focusToggledHeader = () => {
   }
 };
 
-export const focusH3 = (root = document) => {
+export const focusH3 = (index, root) => {
   scrollTo('topContentElement');
-  if (!focusFirstError(root)) {
+  if (!focusFirstError(index, root)) {
     focusElement('#main h3');
   }
 };
 
-export const focusAlertH3 = (root = document) => {
+export const focusAlertH3 = (_index, root = document) => {
   scrollTo('topContentElement');
   const alert = $('va-alert[visible="true"]', root);
   // va-alert header is not in the shadow DOM, but still the content doesn't
